# perf: use constructable stylesheets