--- a/packages/integration-karma/test/shadow-dom/multiple-templates/x/multi/a.css
+++ b/packages/integration-karma/test/shadow-dom/multiple-templates/x/multi/a.css
@@ -1,0 +1,3 @@
+.blue {
+    color: blue;
+}