--- a/packages/integration-karma/test/shadow-dom/multiple-templates/x/multi/a.html
+++ b/packages/integration-karma/test/shadow-dom/multiple-templates/x/multi/a.html
@@ -1,0 +1,3 @@
+<template>
+  <div class="blue">a</div>
+</template>