--- a/packages/integration-karma/test/shadow-dom/multiple-templates/x/multi/b.css
+++ b/packages/integration-karma/test/shadow-dom/multiple-templates/x/multi/b.css
@@ -1,0 +1,3 @@
+.red {
+    color: red;
+}