--- a/packages/integration-karma/test/shadow-dom/multiple-templates/x/multi/b.html
+++ b/packages/integration-karma/test/shadow-dom/multiple-templates/x/multi/b.html
@@ -1,0 +1,3 @@
+<template>
+  <div class="red">b</div>
+</template>