--- a/packages/integration-karma/test/shadow-dom/multiple-templates/x/multi/multi.js
+++ b/packages/integration-karma/test/shadow-dom/multiple-templates/x/multi/multi.js
@@ -1,0 +1,16 @@
+import { LightningElement, api } from 'lwc';
+import A from './a.html';
+import B from './b.html';
+
+export default class Multi extends LightningElement {
+    current = A;
+
+    @api
+    next() {
+        this.current = this.current === A ? B : A;
+    }
+
+    render() {
+        return this.current;
+    }
+}