--- a/packages/perf-benchmarks-components/scripts/generate-styled-components.js
+++ b/packages/perf-benchmarks-components/scripts/generate-styled-components.js
@@ -45,7 +45,7 @@ export function generateStyledComponents() {
     const allComponentsFilename = path.join(tmpDir, 'src/benchmark/styledComponents.js');
     const allComponents = `
         ${components.map((mod, i) => `import cmp${i} from ${JSON.stringify(mod)}`).join(';')};
-        export default [${components.map((_, i) => `cmp${i}`).join(',')}];
+        export default [${components.map((_, i) => `cmp${i}`).join(',')}];`;
     fs.writeFileSync(allComponentsFilename, allComponents, 'utf-8');
 
     return {