--- a/packages/perf-benchmarks/src/__benchmarks__/engine-dom/benchmark-styled-component/ss-styled-component-create-1k-same.benchmark.js
+++ b/packages/perf-benchmarks/src/__benchmarks__/engine-dom/benchmark-styled-component/ss-styled-component-create-1k-same.benchmark.js
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) 2018, salesforce.com, inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: MIT
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
+ */
+
+import StyledComponent from 'perf-benchmarks-components/dist/dom/benchmark/styledComponent.js';
+import { styledComponentBenchmark } from '../../../utils/styledComponentBenchmark';
+
+// Create 1k components with the same CSS in each component
+styledComponentBenchmark(`ss-benchmark-styled-component/create/1k/same`, StyledComponent);