## Details

After looking into [the synthetic-to-native shadow migration](https://salesforce.quip.com/sYZaADXA4arY), it's clear to me that we need a performant solution for including the same stylesheet in multiple components (e.g. `@import 'shared.css'`). And through manual testing, I saw that constructable stylesheets really are a significant perf improvement in Chrome.

This PR is a revised version of #2290 that aims to be less invasive to the codebase while still bringing perf improvements. I also added some new Tachometer benchmarks to verify the perf gain.

My goal is to improve native shadow performance without regressing synthetic shadow performance. To achieve this, we're only using constructable stylesheets in the per-component level, not at the global level. (See [this doc](https://salesforce.quip.com/eUqbA9RVvBfs) for why that matters.)

<details><summary>Benchmark results</summary>

|Benchmark| Browser | Before | After | Delta |
| --- | --- | --- | --- | --- |
|dom-ss-styled-component-create-1k-different | chrome | 334ms-337ms | 335ms-338ms | -0.2% - 0.8% **(~same)** |
|dom-ss-styled-component-create-1k-different | firefox | 296ms-298ms | 297ms-299ms | -0.2% - 0.9% **(~same)** |
|dom-ss-styled-component-create-1k-different | safari | 161ms-163ms | 161ms-163ms | -0.8% - 0.9% **(~same)** |
|dom-ss-styled-component-create-1k-same | chrome | 105ms-107ms | 105ms-107ms | -0.6% - 1.0% **(~same)** |
|dom-ss-styled-component-create-1k-same | firefox | 87ms-89ms | 87ms-89ms | -1.0% - 0.9% **(~same)** |
|dom-ss-styled-component-create-1k-same | safari | 54ms-55ms | 54ms-54ms | -0.9% - 0.9% **(~same)** |
|dom-styled-component-create-1k-different | chrome | 330ms-332ms | 294ms-295ms | -11.5% - -10.6% **(faster)** |
|dom-styled-component-create-1k-different | firefox | 182ms-187ms | 173ms-177ms | -7.0% - -3.6% **(faster)** |
|dom-styled-component-create-1k-different | safari | 150ms-158ms | 146ms-149ms | -6.8% - -1.7% **(faster)** |
|dom-styled-component-create-1k-same | chrome | 148ms-150ms | 101ms-103ms | -32.0% - -30.9% **(faster)** |
|dom-styled-component-create-1k-same | firefox | 99ms-103ms | 83ms-87ms | -18.6% - -13.5% **(faster)** |
|dom-styled-component-create-1k-same | safari | 63ms-64ms | 55ms-56ms | -13.4% - -11.5% **(faster)** |

</details>

(The `ss-` ones use synthetic shadow.) Basically, the native shadow ones are massively improved (especially in Chrome), but the synthetic shadow ones are the same.

## Does this PR introduce breaking changes?

* ✅ `No, it does not introduce breaking changes.`

There is a small change in how light DOM components work. Previously, if you switched templates on a light DOM component, then the old `<style>` would be removed from the DOM and replaced with the new `<style>`. Now the new `<style>` is merely added, while the old `<style>` remains. I don't think this will be a big problem in practice.

## GUS work item
W-9125079