@nolanlawson 2021-08-31
@jodarove @divmain Do you foresee any problems with this PR and rehydration? My assumption was:

1. SSR'd content would use regular `<style>`s (until [declarative constructable stylesheets](https://github.com/WICG/webcomponents/issues/939) ship)
2. During hydration, the client will remove the `<style>` and replace it with a constructable stylesheet (or keep it, I dunno, depends on perf/FOUC implications).